refactor(P4-2): hooks を use-<name>.ts へ統一（Batch1）
- useNotifications → use-notifications / useNotificationPreferences →
  use-notification-preferences / usePushNotifications → use-push-notifications
- 空・未参照の useSSEConnection.ts を削除
- import パスとテストファイル名を更新

本番型チェッククリーン。
