+++ src/components/notification/NotificationCenter.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useRef, useEffect } from "react";
 import { Bell, Settings, CheckCheck, Send } from "lucide-react";
-import { useNotifications } from "@/hooks/useNotifications";
+import { useNotifications } from "@/hooks/use-notifications";
 import { NotificationList } from "./NotificationList";
 import { Button } from "@/components/ui/button";
 import {
