+++ src/hooks/use-notification-preferences.ts
